Separate the class hierarchies for Fusion IR and Kernel IR (#428)
This PR introduces a hard split between the Fusion IR and the Kernel IR: each form has a dedicated class hierarchy. This means that we're free to specialize and evolve each IR without having to worry about the internal details of the "other side".

Separate class hierarchies also make the C++ static type system work for us, accidental mixes would be detected early, at compile time.

The PR touches a lot of code since the new types triggered a cascading set of changes. A lot of the changes are simple, but there are a few notable differences:

- The Kernel IR is owned by the Kernel object, and with a few minor details (kir::TensorView::fuserTv) it is largely decoupled from the Fusion IR

- After the initial lowering pass (LoopNestGenerator::loweredExprs), everything is Kernel IR

- No more `TensorView::unsafeClone(). Replaced with a bit smaller hack.

- Dedicated Kernel IR visitor (kir::IrVisitor)

- There's a dedicated expression evaluator for the Kernel IR (kir::ExpressionEvaluator)

- GpuLower::lowerExpr() can be used to automatically lower a Fusion IR expression node