Specify both this and relative positions when setting computeAt (#482)
* Specify both this and relative positions when setting computeAt.

The problem is that TensorView::setThisComputeAtAxis may return a wrong
position when a new broadcast axis is added to a TensorDomain that already
has broadcast axes. It may not correctly disambiguate which broadcast
axes of the producer should match with one in the consumer. Root mapping
had a similar problem, which was resolved by using the broadcast flag
list, but that list is only applicable to root domains.

Instead of fixing setThisComputeAtAxis, this commit changes all the
uses of setComputeAt to pass both the this and relative positions so
that setThisComputeAtAxis is no longer needed. As far as I see, each
use case of setComputeAt does know the value of this position (some
use cases needed minor changes), so all of them are replaced with the
another interface of setComputeAt.

Fixes #477