Refactor lowering (#597)
Redesigned the lowering steps based on a new mapping concept. 

`ComputeAtMap` is the new foundation of almost all the lowering steps. It has three variants, `LOOP`, `PARALLEL`, and `INDEX`. Each map captures an equivalence relationship among `IterDomain`s in the context of loop nesting, parallelization, or indexing. For example, when an `IterDomain` is mapped with another `IterDomain` in the `LOOP` `ComputeAtMap`, they should correspond to the same `kir::ForLoop` object. Similarly, when an `IterDomain` is mapped in the `INDEX` loop, it should use the same indexing expression.

With this PR, different concepts of mappings are explicitly captured in different mappings, whereas previously they were implicitly embedded in various parts of the code. Hopefully, it should make the lowering more robust and easier to extend further.

Co-authored-by: Naoya Maruyama <nmaruyama@nvidia.com>