Adds support of broadcast and reduction with shift (#870)
Adds support of broadcast and reduction.

Additionally, refactored how kir::Predicate is created. For exprs requiring block sync, the predicate is created at the indexing step, but that doesn't need to be, so it's done at the same time as other exprs. This also makes PredicateType::InternalSync unnecessary, so removed it.