Fix an issue with broadcast mapping (#1024)
In ComputeAtRootDomainMap, a mapping table from each broadcast axis to
concrete axes is built. The mapping is first initialized with fusion
output tensors because they are the final concrete axes for broadcasts,
and then traverse a fusion backward.

There's a logic bug with multi-output expressions (i.e., welford) that
happens when a broadcast axis is used with a welford expression, and if an
output of the op is not used. The mapping for the axis gets never
initialized as its tensor is not an output, which invalidates the
overall traversal logic. This PR makes sure all outputs of multi-output
expressions are initialized properly.