Partial split (#1068)
* Shift without padding

Adds a boolean option to `shift` to disable padding. Shifting without
padding only sets values in a range that correspond to the valid range
in the input tensor. For example, suppose a tensor is defined from 0 to
N, shifting it to right by one produces an output that is defined only
from 1 to N. Similarly, if it is shifted to left, the output is defined
from 0 to N-1. Only valid range is adjusted, and the overall extent is
kept. IterDomain::start_ and IterDomain::stop_ define the valid range of
a domain.

Valid range is only defined in root domains. When a "partial" domain is
merged or split, the output domain does not inherit the range
information directly as it is not possible to define the valid range
just using two values. As such, loops still start from 0 to
IterDomain::extent. Predicates are used to enforce the valid range
defined in root domains.

It was necessary to modify blockReduce (and blockWelford) to support
non-zero start values since they gather the final result to thread 0
even when it might be disabled to run by read_write_pred.  When start is
not zero, the final output, which is held by thread 0, is not written to
the output tensor. This is fixed by having two predicates: one for reads
and another for writes. The read predicate is the same as before, but
the write predicate only looks at non-reducation axes so that the final
output is properly stored by thread 0. This separation of predicates is
only done when reduction axes have non-zero start values.