Bn fp16 io alias update (#1207)
This PR allows us to handle BN inplace update on running stats with reduced precision.

Before this PR, we are blindly aliasing computed running stats, assuming the input would be a direct input to fusion. This assumption is violated when model is casted to reduced precision, where fusion parser would cast them to float32.
The PR accounts for that case, it does two things:

traces the data flow of BN running stats back to the fusion input for aliasing;
casts intermediate running stats to the dtype of fusion input, so we can actually share the same buffer to mimic in-place update.