Make the reduction work buffer volatile (#1301)
It's used as a volatile pointer when writing into it, but not reading
it. This commits changes it to be always volatile.

I started to see temporary validation failures with FusionWelfordShmoo and FusionGridReduction3dim1.
I initially thought there may be some bug in some of the recent changes
around the grid synchronization, however, I don't see any potential
race condition, except for this missing volatile usage, which has been
the case ever since gridReduce was added. I suspect it just hasn't been
exposed until the recent changes.