Casted alias (#1480)
supports alias on casted inputs. This allows us to handle fp16 running stats (fp16 norm layers).

Arguably we would want to do this in the parser code instead of inside Fusion::aliasOutputToInput. But that means we'll duplicate this code in a few places, which is ugly as well.

TODO:
properly handle segment io alias cpp test in PR #1471