consistent clamp for reversed min/max bound (#1572)
Fixes pytorch#75558 (comment)

Updated clamp logic to be consistent with aten. This avoids us producing different result when clamp was given a min/max argument where min > max.
We don't have an issue opened for this. It is also tricky to argue the right behavior, but getting better consistency with eager is always(?!) a good thing.