TransformPropagator refactor (#1769)
This refactors `TransformPropagator` as a 3-level abstraction: `TransformPropagator` is a subclass of `MaxRootDomainInfoPropagator`, while `MaxRootDomainInfoPropagator` is a subclass of `MaxInfoPropagator`.

`MaxInfoPropagator` implements the Dijkstra algorithm for propagating on the DAG, but it has no knowledge what "information" we are trying to preserve and what we are propagating.

`MaxRootDomainInfoPropagator` inherits `MaxInfoPropagator` for preserving most root/rfactor domain information. But it does not have any knowledge about what we are propagating.

`TransformPropagator` further inherits `MaxRootDomainInfoPropagator` and propagates transformations of leaf IDs.