Make inlining even more modular (#2004)
I don't like `TensorView::setComputeAt`  and `TensorView::setMaxProducer`, they are private, and I can not use them conveniently. It would be better if there is some public method of `TensorView` that allows directly setting the CA position of a TV with necessary validation. So I added two public methods `TensorView::inlineAt` and `TensorView::updateMaxProducerPosition` and removed  `TensorView::setComputeAt`  and `TensorView::setMaxProducer`.

The `inlineAt` can be safely used publicly. It will not inline into disallowed dimensions, and the max producer position will be kept consistent. There are two ways of using `inlineAt`: 

If you only want to set the CA position of a single tensor, then simply do 
```C++
tv->inlineAt(pos, /*best_effort=*/true);
```

If you want to set the CA position of multiple tensors, then you can do
```C++
MaxPosCalculator calc;
for (auto tv : tensors) {
  tv->inlineAt(pos, /*best_effort=*/true, &calc);
}
```

In both case, the max producer position will be updated at the end of the `inlineAt` call. Manually constructing the object of `MaxPosCalculator` is mainly for performance reasons: we don't want to build unmappable dimensions every time we call `inlineAt`. If we want to inline multiple tensors, we should build at the beginning and use it in all `inlineAt` calls.

Even though `inlineAt` always updates the max producer position automatically, there are still cases where we want to manually trigger an update of the max producer position, and the `updateMaxProducerPosition` is designed for such a purpose. It is mainly used for grouped reductions.

**With `inlineAt`, I can refactor inlining to make it even more modular:**

There is no longer an `InlinePropagator`. Innermost inlining is now just a dumb for loop:

```C++
MaxPosCalculator calc;
for (auto tv : all_tvs) {
  tv->inlineAt(-1, /*best_effort=*/true, &calc);
}
```

For standard and best effort inlining, we need first to do a propagation to find the positions in each tensor mapped to the given reference tensor's given position. With the positions calculated, inlining is again a dumb for loop.